exercicio-2: Multiplicação e potenciação em uma única função Fixes #6
