+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "files.associations": {
+        "exercicio-2-funcoes.h": "c"
+    }
+}
