+++ Lista-1/exercicio-2/exercicio-2-funcoes.h
@@ -1,0 +1,9 @@
+#ifndef EXERCICIO_2_FUNCOES_H
+#define EXERCICIO_2_FUNCOES_H
+
+void impressaoMatriz(float matriz[3][3]);
+void copiaMatriz(float matrizA[3][3], float matrizB[3][3]);
+void multiplicacaoMatrizes(float matrizA[3][3], float matrizB[3][3], float matrizC[3][3]);
+void elevaMatriz(float matrizA[3][3], float matrizB[3][3], int n);
+
+#endif //EXERCICIO_2_FUNCOES_H
