+++ Lista-1/exercicio-2/exercicio-2.c
@@ -1,10 +1,7 @@
 #include <stdio.h>
+#include "exercicio-2-funcoes.h"
 
-//Declaração das funções:
-void impressaoMatriz(float matriz[3][3]);
-void copiaMatriz(float matrizA[3][3], float matrizB[3][3]);
-void multiplicacaoMatrizes(float matrizA[3][3], float matrizB[3][3], float matrizC[3][3]);
-void elevaMatriz(float matrizA[3][3], float matrizB[3][3], int n);
+void multiplica(float matrizA[3][3], float matrizB[3][3], float matrizC[3][3], int n);
 
 int main()
 {
@@ -21,92 +18,28 @@ int main()
         {0.3, 0.1, 0.9}
     };
 
+    //Declaração do expoente n:
+    int n = 3; 
+
     //Declaração da matriz C de resultado:
-    float C[3][3];
+    float C[3][3]; 
 
-    //Declaração dos indices:
-    int i, j, k;
+    multiplica(A, B, C, n); //C = (A * B)^n
 
-    //Multiplicação das matrizes A e B, resultado em C:
-    multiplicacaoMatrizes(A, B, C);    
-
-    //Impressão das matrizes:
-    printf("Matriz A:\n");
-    impressaoMatriz(A);
-
-    printf("\nMatriz B:\n");
-    impressaoMatriz(B);
-
+    //Impressão da matriz C:
     printf("\nMatriz C:\n");
     impressaoMatriz(C);
-
-    //Recebe expoenente n:
-    int n;
-    printf("\nDigite o valor de n: ");
-    scanf("%d", &n);
-
-    //Declaração da matriz C^n:
-    float Cn[3][3];
-    elevaMatriz(C, Cn, n);
-    
-
-    //Impressão da matriz C^n:
-    printf("\nMatriz C^n:\n");
-    impressaoMatriz(Cn);
 
     return 0;
 }
 
-void impressaoMatriz(float matriz[3][3]) {
-    int i, j;
+void multiplica(float matrizA[3][3], float matrizB[3][3], float matrizC[3][3], int n) {
+    //Primeiro multiplica as matrizes A e B, armazenando o resultado na matriz C: 
+    multiplicacaoMatrizes(matrizA, matrizB, matrizC);
 
-    //Imprime cada linha da matriz:
-    for (i = 0; i < 3; i++) {
-        printf("[");
-        for (j = 0; j < 3; j++) {
-            printf("%.6f", matriz[i][j]);
-            if (j < 2) {
-                printf(", ");
-            }
-        }
-        printf("]\n");
-    }
-}
+    float matrizAux[3][3];
 
-void copiaMatriz(float matrizA[3][3], float matrizB[3][3]) {
-    int i, j;
-
-    //Copia cada elemento da matriz A para a matriz B:
-    for (i = 0; i < 3; i++) {
-        for (j = 0; j < 3;j++) {
-            matrizB[i][j] = matrizA[i][j];
-        }
-    }
-}
-
-void multiplicacaoMatrizes(float matrizA[3][3], float matrizB[3][3], float matrizC[3][3]) {
-    int i, j, k;
-
-    //Soma o produto dos elementos de cada linha de A com cada coluna de B:
-    for (i = 0; i < 3; i++) {
-        for (j = 0; j < 3; j++){
-            matrizC[i][j] = 0;
-            for (k = 0; k < 3; k++) {
-                matrizC[i][j] += matrizA[i][k] * matrizB[k][j];
-            }
-        }
-    }
-}
-
-void elevaMatriz(float matrizA[3][3], float matrizB[3][3], int n) {
-    int i, j, k;
-    float matriz[3][3];
-    copiaMatriz(matrizA, matriz);
-    copiaMatriz(matrizA, matrizB);
-
-    //Multiplica a matriz A por si mesma n vezes:
-    for (i = 1; i < n; i++) {
-        multiplicacaoMatrizes(matrizA, matriz, matrizB);
-        copiaMatriz(matrizB, matriz);
-    }
+    //Depois eleva a matriz C ao expoente n:
+    elevaMatriz(matrizC, matrizAux, n);
+    copiaMatriz(matrizAux, matrizC);
 }
